refactor(library/HTTP): renames "RequestError.ts" to "Error/Request.ts"
